backend/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
python-dotenv==1.0.1
scikit-learn==1.8.0
joblib==1.5.3
numpy==2.3.2
pydantic==2.11.7
bcrypt==3.2.0
cryptography==46.0.3
python-multipart==0.0.9
pandas==2.3.2
